// src/app/models/tournament-player.model.ts
export class TournamentPlayer {
  tournamentId?: any;
  playerId?: any;
  playerNumber?: number;
  place?: number;
  gamePoints?: number;
  tournamentPoints?: number;
  competitionPoints?: number;
  averageHand?: number;
  maxHand?: number;
  maxHandDealId?: any;
  averagePointsTaken?: number;
  handsTaken?: number;
  selvesTaken?: number;
  passiveLosses?: number;
  feeds?: number;
  averageFeed?: number;
  maxFeed?: number;
  otherSelvesLosses?: number;
  averageOSLoss?: number;
  maxOSLoss?: number;
  drawsAmt?: number;
  maxWinRow?: number;
}